pyproject.toml
[tool.poetry]
name = "project_harvey"
version = "0.1.0"
description = ""
# FIX: Use the simple string format for authors
authors = ["nathanmendis <nathanmendis17@gmail.com>"] 
readme = "README.md"
packages = [{include = "project_harvey"}]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = ">=5.2.7,<6.0.0"
openai = ">=2.3.0,<3.0.0"
pydantic = ">=2.12.0,<3.0.0"
httpx = ">=0.28.1,<0.29.0"
"python-dotenv" = ">=1.1.1,<2.0.0"
tqdm = ">=4.67.1,<5.0.0"
"django-tailwind-cli" = ">=4.4.2,<5.0.0"
"django-tailwind" = ">=4.2.0,<5.0.0"

channels = ">=4.3.1,<5.0.0"
daphne = ">=4.2.1,<5.0.0"
"django-redis" = ">=6.0.0,<7.0.0"

langchain = "^1.0.7"
langchain-google-genai = "^3.0.3"
langchain-core = "^1.0.5"
langchain-community = "^0.4.1"
langgraph = ">=0.2.0"
langgraph-checkpoint = "^3.0.1"
langgraph-checkpoint-sqlite = ">=0.1.0"

langchain-huggingface = "^1.1.0"
django-browser-reload = "^1.21.0"
pdfminer-six = "^20251107"
python-docx = "^1.2.0"
beautifulsoup4 = "^4.14.3"
requests = "^2.32.5"
djangorestframework = "^3.16.1"
langchain-postgres = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.3.1"}
pypdf = "^6.4.0"

langchain-groq = "^1.1.1"
sentence-transformers = "^5.2.2"
google-auth = "^2.48.0"
google-auth-oauthlib = "^1.2.4"
google-auth-httplib2 = "^0.3.0"
google-api-python-client = "^2.188.0"
cryptography = "^46.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = []
